utils/py/requirements.txt
PyYAML>=6.0
aiofiles>=0.8.0
requests>=2.25.0 